move geometry code to central com package
