+++ packages/rise/rise/lib/location.py
@@ -4,6 +4,7 @@
 from datetime import datetime
 import logging
 from typing import Literal, Optional, assert_never
+from com.helpers import parse_bbox, parse_date, parse_z
 import geojson_pydantic
 from pydantic import BaseModel, field_validator
 import shapely
@@ -13,9 +14,6 @@ from com.geojson.types import GeojsonFeatureCollectionDict
 from rise.lib.helpers import (
     merge_pages,
     no_duplicates_in_pages,
-    parse_bbox,
-    parse_date,
-    parse_z,
 )
 from rise.lib.types.helpers import ZType
 from rise.lib.types.includes import LocationIncluded
