+++ packages/rise/rise/lib/location_test.py
@@ -1,9 +1,10 @@
 # Copyright 2025 Lincoln Institute of Land Policy
 # SPDX-License-Identifier: MIT
 
+from com.helpers import await_
 import pytest
 from rise.lib.cache import RISECache
-from rise.lib.helpers import flatten_values, getResultUrlFromCatalogUrl, await_
+from rise.lib.helpers import flatten_values, getResultUrlFromCatalogUrl
 from rise.lib.location import LocationResponseWithIncluded
 from rise.rise_edr import RiseEDRProvider
